rearranged coordinate function
