--- src/DetailModal.js
@@ -1,58 +1,0 @@
-import React, { useState, useEffect } from 'react';
-import ApiManager from './modules';
-
-const DetailModal = props => {
-    const [details, setDetails] = useState([])
-
-    const getDetails = e => {
-        const results = []
-        ApiManager.getList("categorymembers", "max", "Category:Lists_of_shipwrecks_by_location")
-            .then(response => {
-                console.log("response 1", response)
-                response.query.categorymembers.forEach(category => {
-                    ApiManager.getGenerator("links", "info", category.pageid, "max")
-                        .then(response => {
-                            console.log("response 2", response)
-                            if (response.query) {
-                                Object.entries(response.query.pages).forEach(entry => {
-                                    console.log('entry', entry)
-                                    if (entry[1].length) {
-                                        ApiManager.getArticleByProp("coordinates", entry[0])
-                                        .then(response => {
-                                            console.log("response 3", response)
-                                            Object.entries(response.query.pages).forEach(entry => {
-                                                console.log("entry2", entry)
-                                                if (entry[1].coordinates) {
-                                                    console.log("got the coords")
-                                                    results.push({
-                                                        "pageid": entry[0],
-                                                        "coordinates": {
-                                                            "lat": entry[1].coordinates[0].lat,
-                                                            "lon": entry[1].coordinates[0].lon
-                                                        }
-                                                    })
-                                                    console.log("results prog", results)
-                                                }
-                                            })
-                                            setDetails(results)
-                                        })
-                                    }
-                                })
-                            }
-                        })
-                    })
-                })
-    }
-
-    useEffect(() => {
-        getDetails()
-    }, [])
-
-    return (
-        <>
-            <h3>Details</h3>
-        </>
-    )
-}
-
-export default DetailModal
